requirements.txt
requests
beautifulsoup4
camelot-py
opencv-python